+++ InventoryControl/GlobalUsings.cs
@@ -9,6 +9,7 @@ global using InventoryControl.Domain.Identity;
 global using InventoryControl.Domain.Integrations;
 global using InventoryControl.Domain.Orders;
 global using InventoryControl.Domain.Products;
+global using InventoryControl.Domain.Products.Events;
 global using InventoryControl.Domain.Shared;
 global using InventoryControl.Domain.Stock;
 
@@ -29,6 +30,8 @@ global using InventoryControl.Infrastructure;
 global using InventoryControl.Infrastructure.Auth;
 global using InventoryControl.Infrastructure.Backup;
 global using InventoryControl.Infrastructure.BackgroundJobs;
+global using InventoryControl.Infrastructure.Email;
+global using InventoryControl.Infrastructure.Events;
 global using InventoryControl.Infrastructure.Integrations;
 global using InventoryControl.Infrastructure.Integrations.Abstractions;
 global using InventoryControl.Infrastructure.Integrations.Nuvemshop;